requirements.txt
mysql-connector-python==8.0.20
numpy==1.26.4
pandas==2.2.2
pillow==10.3.0
protobuf==5.26.1
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tzdata==2024.1
